fix(codebase): refactored and cleaned up attestor #4
