+++ src/contractDepositAttestor.spec.ts
@@ -1,0 +1,98 @@
+import { Logger, LogPrinter } from '@nori-zk/proof-conversion';
+import {
+    ContractDepositAttestorInput,
+    ContractDepositAttestor,
+    buildContractDepositLeaves,
+    getContractDepositWitness,
+    ContractDeposit,
+} from './contractDepositAttestor.js';
+import { sp1ConsensusMPTPlonkProof } from './test_examples/sp1-mpt-proof/sp1ProofMessage.js';
+import { Bytes20, Bytes32 } from './types.js';
+import {
+    computeMerkleTreeDepthAndSize,
+    foldMerkleLeft,
+    getMerkleZeros,
+} from './merkle/merkleTree.js';
+import { UInt64 } from 'o1js';
+
+const logger = new Logger('ContractDepositAttestor');
+new LogPrinter('[TestEthProcessor]', [
+    'log',
+    'info',
+    'warn',
+    'error',
+    'debug',
+    'fatal',
+    'verbose',
+]);
+
+describe('Contract Storage Slot Deposit Attestor Test', () => {
+    test('pipeline', async () => {
+        // Analyse zk program
+        const contractDepositAttestorAnalysis =
+            await ContractDepositAttestor.analyzeMethods();
+        logger.log(
+            `ContractDepositAttestor analyze methods gates length '${contractDepositAttestorAnalysis.compute.gates.length}'.`
+        );
+
+        // Build zk program
+        const { verificationKey } = await ContractDepositAttestor.compile({
+            forceRecompile: true,
+        });
+        logger.log(
+            `ContractDepositAttestor contract compiled vk: '${verificationKey.hash}'.`
+        );
+
+        // Build contractStorageSlot from sp1 mpt message.
+        const contractStorageSlots =
+            sp1ConsensusMPTPlonkProof.contract_storage_slots.map((slot) => {
+                return new ContractDeposit({
+                    address: Bytes20.fromHex(slot.slot_key_address.slice(2)),
+                    value: Bytes32.fromHex(slot.value.slice(2)),
+                });
+            });
+
+        // Build leaves
+        const leaves = buildContractDepositLeaves(contractStorageSlots);
+
+        // Pick an index
+        let index = sp1ConsensusMPTPlonkProof.contract_storage_slots.length - 1;
+
+        // Find Value
+        const slotToFind = contractStorageSlots.find((_, idx) => idx === index);
+
+        if (!slotToFind) throw new Error(`Slot at ${index} not found`);
+
+        // Compute path
+        const path = getContractDepositWitness([...leaves], index);
+
+        // Compute root
+        const { depth, paddedSize } = computeMerkleTreeDepthAndSize(
+            leaves.length
+        );
+        const rootHash = foldMerkleLeft(
+            leaves,
+            paddedSize,
+            depth,
+            getMerkleZeros(depth)
+        );
+
+        // Build ZK input
+        const input = new ContractDepositAttestorInput({
+            rootHash,
+            path,
+            index: UInt64.from(index),
+            value: slotToFind,
+        });
+
+        logger.log(`Generated input ${JSON.stringify(input)}`);
+
+        // Prove deposit with sample data.
+        let start = Date.now();
+        const output = await ContractDepositAttestor.compute(input);
+        let durationMs = Date.now() - start;
+        logger.log(`ContractDepositAttestor.compute took ${durationMs}ms`);
+
+        expect(output.proof.publicOutput.toBigInt()).toBe(rootHash.toBigInt());
+    });
+});
