+++ src/test_examples/sp1-mpt-proof/sp1ProofMessage.ts
