+++ src/utils.spec.ts
@@ -1,5 +1,5 @@
 import { decodeConsensusMptProofNew } from './utils';
-import { sp1ConsensusMPTPlonkProof } from './test_examples/sp1-with-mpt/sp1ProofMessage.js';
+import { sp1ConsensusMPTPlonkProof } from './test_examples/sp1-mpt-proof/sp1ProofMessage.js';
 
 describe('ConsensusMPT marshaller Integration Test', () => {
     test('should decode consensus mpt proof', async () => {
